Added select all /single getting ready for delete
Brought back media-options.  Not sure where that was going.  My way may
not be the right way but it gets the job done.  Select single and select
all are working as well as added number of total attachments with number
of attachments available on screen for bulk operations.  Delete here we
come!