Delete is functional!
Delete works both per item and in bulk.  Brought back the media-options
pop-up to display number of items selected and bulk delete,  bulk
tagging (not working yet, will add functionality later) and unselect.
Chose not to show small preview of selected items.  Thought it looked
cleaner this way but it could easily be added back in.