Cleaned up Ajax calls
Placed ajax calls for deleting items into it's own function and added
another ajax call function that calls all media items for when the user
checks all and deletes all so they are not left with a blank screen.
This new function will come in handy when I start changing how the input
box works (Media Tags) and for a separate link just to start over.