Filter files by size.
Perhaps we can consider saving filesize as post meta, so querying
could go through a post_meta search.

To shoehorn this in, add a custom version of wp_ajax_query_attachments
so in our wp_prepare_attachment_for_js callback any files that don't
match the filesize query get taken out of the query results.

This is a janky implementation since we're filtering results after
the query has run, but works as a hobbling prototype.